refactor(FR-2529): extract backend.ai-client as standalone workspace package
Extract src/lib/backend.ai-client-esm.ts (6015 lines) into packages/backend.ai-client/
as a proper workspace package with tsup build.

- Split monolithic file into 28 modular resource class files
- Build with tsup: ESM output, .d.ts auto-generation, sourcemaps
- Update consumers: craco alias, global-stores import, jest mock
- Replace original file with thin re-export shim
- Follow backend.ai-ui patterns: naming, versioning, license

Resolves #6612(FR-2529)
