+++ packages/backend.ai-client/src/resources/index.ts
@@ -1,0 +1,26 @@
+export { ResourcePreset } from './resource-preset';
+export { VFolder } from './vfolder';
+export { Agent } from './agent';
+export { AgentSummary } from './agent';
+export { StorageProxy } from './storage-proxy';
+export { Keypair } from './keypair';
+export { ResourcePolicy } from './resource-policy';
+export { ContainerImage } from './container-image';
+export { ComputeSession } from './compute-session';
+export { SessionTemplate } from './session-template';
+export { ModelService } from './model-service';
+export { Resources } from './resources';
+export { Group } from './group';
+export { Domain } from './domain';
+export { Maintenance } from './maintenance';
+export { User } from './user';
+export { ScalingGroup } from './scaling-group';
+export { Registry } from './registry';
+export { Setting } from './setting';
+export { Service } from './service';
+export { UserConfig } from './user-config';
+export { Enterprise } from './enterprise';
+export { Cloud } from './cloud';
+export { Pipeline, PipelineJob, PipelineTaskInstance } from './pipeline';
+export { EduApp } from './edu-app';
+export { utils } from './utils';
