+++ pnpm-workspace.yaml
@@ -1,6 +1,7 @@
 packages:
   - react
   - packages/backend.ai-ui
+  - packages/backend.ai-client
   - packages/eslint-config-bai
   - packages/backend.ai-docs-toolkit
   - packages/backend.ai-webui-docs
