+++ react/__test__/backendAiClient.mock.js
@@ -1,8 +1,10 @@
-// Mock for backend.ai-client-esm (resolved via webpack alias in real builds)
+// Mock for backend.ai-client package
 class MockClient {}
 class MockClientConfig {}
 
 module.exports = {
+  Client: MockClient,
+  ClientConfig: MockClientConfig,
   backend: {
     Client: MockClient,
     ClientConfig: MockClientConfig,
