+++ react/src/react-app-env.d.ts
@@ -8,11 +8,6 @@
 //   export { graphql as default } from 'react-relay';
 // }
 
-declare module 'backend.ai-client-esm' {
-  const ai: { backend: { Client: any; ClientConfig: any } };
-  export = ai;
-}
-
 type ArrayElement<ArrayType extends readonly unknown[]> =
   ArrayType extends readonly (infer ElementType)[] ? ElementType : never;
 type ArgumentTypes<F extends Function> = F extends (...args: infer A) => any
